+++ pkg/middleware/mirror/mirror.go
@@ -15,7 +15,7 @@ import (
 
 // Init registers the mirror middleware.
 func Init() error {
-	return middleware.RegisterTyped([]string{"mirror"}, func(opts Options) (app.HandlerFunc, error) {
+	return middleware.Register([]string{"mirror"}, func(opts Options) (app.HandlerFunc, error) {
 		if opts.ServiceID == "" {
 			return nil, errors.New("mirror: service_ID cannot be empty")
 		}